new styles for insights on feed
