+++ src/components/Insight/InsightCardInternals.js
@@ -12,13 +12,6 @@ import styles from './InsightCard.module.scss'
 import { DesktopOnly } from '../Responsive'
 import { SignalTypeIcon } from '../SignalCard/controls/SignalControls'
 
-const AWAITING_APPROVAL_STATE = 'awaiting_approval'
-const AwaitingApproval = () => (
-  <div className={styles.awaiting}>
-    <Icon type='awaiting' className={styles.awaiting__icon} /> Awaiting approval
-  </div>
-)
-
 export const makeLinkToInsight = (id, title) => {
   return `https://insights.santiment.net/read/${getSEOLinkFromIdAndTitle(
     id,
@@ -31,9 +24,6 @@ const InsightCardInternals = ({
   user: { id: authorId, username: authorName, avatarUrl },
   title,
   tags,
-  createdAt,
-  publishedAt,
-  state,
   votes: { totalVotes },
   commentsCount,
   votedAt,
@@ -41,9 +31,9 @@ const InsightCardInternals = ({
   withAuthorPic,
   disabled,
   isDesktop,
-  showIcon = false
+  showIcon = false,
+  children
 }) => {
-  const { DD, MMM, YYYY } = getDateFormats(new Date(publishedAt || createdAt))
   const linkToInsight = makeLinkToInsight(id, title)
 
   return (
@@ -54,49 +44,43 @@ const InsightCardInternals = ({
         </DesktopOnly>
       )}
       <div className={cx(styles.main, showIcon && styles.withIcon)}>
-        <div className={styles.top}>
-          <a href={linkToInsight} className={styles.title}>
-            <MultilineText maxLines={2} id='insightCardTitle' text={title} />
-          </a>
+        <div className={styles.description}>
+          <div className={styles.top}>
+            <a href={linkToInsight} className={styles.title}>
+              <MultilineText maxLines={2} id='insightCardTitle' text={title} />
+            </a>
+            <div className={styles.profile}>
+              <ProfileInfo
+                withPic={withAuthorPic}
+                picUrl={avatarUrl}
+                name={
+                  <Link className={styles.name} to={`/profile/${authorId}`}>
+                    {authorName}
+                  </Link>
+                }
+                infoClassName={styles.info}
+              />
+            </div>
+          </div>
+          <div className={styles.chart}>{children}</div>
         </div>
         <div className={styles.bottom}>
-          <div className={styles.profile}>
-            <ProfileInfo
-              withPic={withAuthorPic}
-              picUrl={avatarUrl}
-              name={
-                <Link className={styles.name} to={`/profile/${authorId}`}>
-                  {authorName}
-                </Link>
-              }
-              status={
-                state === AWAITING_APPROVAL_STATE ? (
-                  <AwaitingApproval />
-                ) : (
-                  `${MMM} ${DD}, ${YYYY}`
-                )
-              }
-              infoClassName={styles.info}
-            />
-          </div>
-          <div className={styles.right}>
-            <LikeBtn
-              likesNumber={totalVotes}
-              liked={!!votedAt}
-              onClick={onLike}
-              disabled={disabled}
-              className={styles.likeBtn}
-            />
-            <a
-              href={linkToInsight + '?_wc=1#comments'}
-              className={cx(styles.stat, styles.stat_comments)}
-            >
-              <Icon type='comment' className={styles.commentIcon} />{' '}
-              {commentsCount}
-            </a>
-            <div className={styles.tags}>
-              <InsightTags tags={tags} isDesktop={isDesktop} />
-            </div>
+          <LikeBtn
+            likesNumber={totalVotes}
+            liked={!!votedAt}
+            onClick={onLike}
+            disabled={disabled}
+            className={styles.likeBtn}
+          />
+          <a
+            href={linkToInsight + '?_wc=1#comments'}
+            className={cx(styles.stat, styles.stat_comments)}
+          >
+            <Icon type='comment' className={styles.commentIcon} />{' '}
+            {commentsCount}
+          </a>
+          <div className={styles.tags}>
+            <InsightTags tags={tags} isDesktop={isDesktop} />
           </div>
         </div>
       </div>
