+++ src/components/Insight/InsightCardWithMarketcap.js
@@ -1,28 +1,45 @@
 import React from 'react'
-import { Panel } from '@santiment-network/ui'
+import { Icon, Panel } from '@santiment-network/ui'
 import cx from 'classnames'
 import withSizes from 'react-sizes'
 import InsightCardInternals from './InsightCardInternals'
 import MarketcapChangeWidget from '../PostVisualBacktest'
 import { noTrendTagsFilter } from './utils'
 import { mapSizesToProps } from '../../utils/withSizes'
+import FeedCardDate from '../../pages/feed/GeneralFeed/CardDate/FeedCardDate'
 import styles from './InsightCard.module.scss'
 
+const AWAITING_APPROVAL_STATE = 'awaiting_approval'
+const AwaitingApproval = () => (
+  <div className={styles.awaiting}>
+    <Icon type='awaiting' className={styles.awaiting__icon} /> Awaiting approval
+  </div>
+)
+
 const InsightCard = ({ className, tags, isDesktop, showIcon, ...insight }) => {
-  const { createdAt, updatedAt, publishedAt } = insight
+  const { createdAt, updatedAt, publishedAt, state } = insight
   const filteredTags = tags.filter(noTrendTagsFilter)
+
   return (
     <Panel className={cx(styles.wrapper, styles.wrapper_withMc, className)}>
-      <div className={styles.wrapper_withMc__left}>
-        <InsightCardInternals
-          {...insight}
-          tags={filteredTags}
-          isDesktop={isDesktop}
-          showIcon={showIcon}
-        />
-      </div>
-      {isDesktop && (
+      <InsightCardInternals
+        {...insight}
+        tags={filteredTags}
+        isDesktop={isDesktop}
+        showIcon={showIcon}
+      >
         <div className={styles.wrapper_withMc__right}>
+          <>
+            {state === AWAITING_APPROVAL_STATE ? (
+              <AwaitingApproval />
+            ) : (
+              <FeedCardDate
+                date={publishedAt || createdAt}
+                className={styles.date}
+              />
+            )}
+          </>
+
           <MarketcapChangeWidget
             from={createdAt}
             ticker={(filteredTags[0] || {}).name}
@@ -30,7 +47,7 @@ const InsightCard = ({ className, tags, isDesktop, showIcon, ...insight }) => {
             publishedAt={publishedAt || updatedAt}
           />
         </div>
-      )}
+      </InsightCardInternals>
     </Panel>
   )
 }
