+++ src/components/Insight/ProfileInfo.module.scss
@@ -28,6 +28,7 @@
   flex-direction: column;
   max-width: 100%;
   flex: 1;
+  align-items: center;
 
   &__item:last-of-type {
     margin-bottom: 0;
@@ -52,12 +53,4 @@
   * {
     color: var(--black);
   }
-}
-
-.status {
-  @include text-normal;
-
-  line-height: 1;
-  font-size: 12px;
-  color: var(--waterloo);
 }
