+++ src/components/PostVisualBacktestChart.css
@@ -1,6 +1,6 @@
 .PostVisualBacktestChart {
   min-height: 75px;
-  height: 100px;
-  width: 220px;
+  height: 80px;
+  width: 190px;
   margin: 0 auto;
 }
