+++ src/pages/SonarFeed/ActivityRenderer/ActivityRenderer.module.scss
@@ -96,7 +96,7 @@
   flex-direction: column;
   align-items: center;
   justify-content: center;
-  width: 260px;
+  width: 229px;
 
   @include responsive('phone', 'phone-xs', 'laptop', 'tablet') {
     flex: 1;
