+++ src/pages/feed/GeneralFeed/CardDate/FeedCardDate.module.scss
@@ -3,6 +3,6 @@
 .date {
   @include text('caption');
 
-  color: var(--waterloo);
+  color: var(--casper);
   margin-left: auto;
 }
