+++ src/queries/InsightsGQL.js
@@ -20,6 +20,7 @@ export const INSIGHT_COMMON_FRAGMENT = gql`
       username
       avatarUrl
     }
+    shortDesc
     commentsCount
     __typename
   }
